@@ -11,6 +11,7 @@
     },
     "devDependencies": {
         "@popperjs/core": "^2.10.2",
+        "@vue/compiler-sfc": "^3.2.37",
         "axios": "^0.21",
         "bootstrap": "^5.1.3",
         "laravel-mix": "^6.0.6",
@@ -19,8 +20,12 @@
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.11.1",
+        "vue": "^3.2.37",
+        "vue-loader": "^16.8.3",
         "vue-template-compiler": "^2.6.12"
+    },
+    "dependencies": {
+        "vue-router": "^4.0.13",
+        "vuex": "^4.0.2"
     }
 }
